+++ src/p2p/mod.rs
@@ -2,17 +2,18 @@ mod bootstrap;
 mod broadcast;
 mod connection;
 mod peers;
+mod status;
 mod tor;
 
 pub mod dispatch;
 pub mod server;
-pub mod status;
 
-pub use bootstrap::Bootstrap;
+pub use bootstrap::{Bootstrap, BootstrapState};
 pub use broadcast::Broadcaster;
 pub use connection::{Connection, ConnectionId, Request};
 pub use peers::Peers;
 pub use server::TorConfig;
+pub use status::*;
 
 pub mod message {
     pub use super::broadcast::Broadcast;